fix: organize components in folders
