+++ frontend/src/components/app.tsx
@@ -3,7 +3,7 @@ import Header from './header';
 import SearchBox from './search-box';
 import ResultList from './result-list';
 import { BrowserRouter as Router, Link, Switch, Route } from 'react-router-dom';
-import { UserIcon, SkillIcon } from './svg-icons';
+import { UserIcon, SkillIcon } from './svg-icons/svg-icons';
 import { LoginPage } from './login-page';
 import { MyProfilePage } from './my-profile-page';
 import { ProfilePage } from './profile-page';
