+++ frontend/src/components/svg-icons/svg-icons.tsx
@@ -1,5 +1,8 @@
 import React from 'react';
 
+import './svg-icons.scss';
+
+
 type SvgIconProps = {
   width ?: number;
   height ?: number;
