+++ frontend/src/css/index.scss
@@ -8,4 +8,3 @@
 @import './profile-page';
 @import './skill-page';
 @import './toasts';
-@import './svg-icons';
